+++ test/octo_fetch_test.exs
@@ -10,8 +10,8 @@ defmodule OctoFetchTest do
       download_versions: %{
         "0.3.9" => [
           {:linux, :amd64, "806e1cca4a2a105a36f219a4c212a220569d50a8f13f45f38ebe49e6699ab99f"},
-          {:macos, :amd64, "74599a34dc440c19544f533be2ef14cd4378ec1969b9b4fcfd24158946541869"},
-          {:macos, :arm64, "74599a34dc440c19544f533be2ef14cd4378ec1969b9b4fcfd24158946541869"}
+          {:darwin, :amd64, "74599a34dc440c19544f533be2ef14cd4378ec1969b9b4fcfd24158946541869"},
+          {:darwin, :arm64, "74599a34dc440c19544f533be2ef14cd4378ec1969b9b4fcfd24158946541869"}
         ],
         "0.3.8" => [
           {:linux, :amd64, "530723d95a51ee180e29b8eba9fee8ddafc80a01cab7965290fb6d6fc31381b3"}
@@ -19,7 +19,7 @@ defmodule OctoFetchTest do
       }
 
     @impl true
-    def download_name(version, :macos, _arch), do: "litestream-v#{version}-darwin-amd64.zip"
+    def download_name(version, :darwin, _arch), do: "litestream-v#{version}-darwin-amd64.zip"
     def download_name(version, :linux, arch), do: "litestream-v#{version}-linux-#{arch}.tar.gz"
   end
 
@@ -47,16 +47,16 @@ defmodule OctoFetchTest do
 
   test "Should return an error if an invalid architecture is provided" do
     capture_log(fn ->
-      assert {:error, "Your platform is not supported for the provided version"} =
+      assert {:error, "Your platform is not supported for the provided version (os=darwin, architecture=bad_arch)"} =
                Litestream.Fetcher.download(".", override_architecture: :bad_arch)
-    end) =~ "Your platform is not supported for the provided version"
+    end) =~ "Your platform is not supported for the provided version (os=darwin, architecture=bad_arch)"
   end
 
   test "Should return an error if an invalid OS is provided" do
     capture_log(fn ->
-      assert {:error, "Your platform is not supported for the provided version"} =
+      assert {:error, "Your platform is not supported for the provided version (os=bad_os, architecture=arm64)"} =
                Litestream.Fetcher.download(".", override_operating_system: :bad_os)
-    end) =~ "Your platform is not supported for the provided version"
+    end) =~ "Your platform is not supported for the provided version (os=bad_os, architecture=arm64)"
   end
 
   @tag :tmp_dir