+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Support for FreeBSD
 
+### Changed
+
+- Switched from `:macos` to `:darwin`
+
 ## [0.1.0] - 2022-10-08
 
 ### Added